fix lat lon not being passed to cell on page load
