--- packages/app/src/lib/components/ToggleGroup2.svelte
@@ -1,82 +1,0 @@
-<script lang="ts">
-	import { createToggleGroup, melt } from '@melt-ui/svelte';
-	import { Check } from 'phosphor-svelte';
-	import { mergeCss } from '$utils/utils';
-
-	interface Props {
-		items: string[];
-		orientation?: 'horizontal' | 'vertical';
-		type?: 'single' | 'multiple';
-		selectedItems?: string[];
-		disabledItems?: string[];
-		onItemSelected?: (selected: string[]) => void;
-		class?: string;
-	}
-
-	let { 
-		items, 
-		selectedItems = [], 
-		disabledItems = [],
-		orientation = 'vertical',
-		type = 'multiple',
-		onItemSelected,
-		class: className
-	}: Props = $props();
-
-	let sortedItems = $derived([...items].sort());
-
-	const {
-		elements: { root, item },
-		states: { value }
-	} = createToggleGroup({
-		type,
-		defaultValue: type === 'single' ? selectedItems[0] : selectedItems,
-		orientation: orientation,
-		onValueChange: ({curr, next}) => {
-		  if(onItemSelected) {
-		    onItemSelected(next);
-		  }
-		  return next 
-		}
-	});
-
-	function isDisabled(itemValue: string): boolean {
-		return disabledItems.includes(itemValue);
-	}
-</script>
-
-<div
-	use:melt={$root}
-	class={mergeCss("flex items-start data-[orientation='vertical']:flex-col data-[orientation='horizontal']:flex-row gap-1", className)}
-	aria-label="Toggle selection"
-	role={type === 'single' ? 'radiogroup' : 'group'}
->
-	{#each sortedItems as itemValue (itemValue)}
-		<button
-			class="pb-1 last-child:pb-0 flex items-center gap-2 w-full text-left cursor-pointer transition-colors hover:bg-gray-50 focus:outline-none focus:bg-gray-50 data-[disabled]:cursor-not-allowed data-[disabled]:opacity-50 data-[disabled]:hover:bg-transparent"
-			use:melt={$item({ value: itemValue, disabled: isDisabled(itemValue) })}
-			aria-label="Toggle {itemValue}"
-			role={type === 'single' ? 'radio' : 'checkbox'}
-			aria-checked={$value && $value.includes(itemValue) || (type === 'single' && $value === itemValue)}
-		>
-			<span 
-				class="w-5 h-5 border-2 border-gray-300 rounded flex items-center justify-center bg-white text-blue-600 transition-colors data-[state='on']:border-blue-600 data-[state='on']:bg-blue-600 data-[state='on']:text-white"
-				class:border-gray-200={isDisabled(itemValue)}
-				class:bg-gray-100={isDisabled(itemValue)}
-				class:text-gray-400={isDisabled(itemValue)}
-				aria-hidden="true"
-			>
-				{#if $value && $value.includes(itemValue) || (type === 'single' && $value === itemValue)}
-					<Check size={16} weight="bold" />
-				{/if}
-			</span>
-			<span 
-				class="text-sm text-gray-700 transition-colors select-none"
-				class:text-gray-400={isDisabled(itemValue)}
-				aria-hidden="true"
-			>
-				{itemValue}
-			</span>
-		</button>
-	{/each}
-</div>
